+++ packages/calcite-components/src/components/loader/loader.tsx
@@ -16,6 +16,13 @@ export class Loader implements LocalizedComponent {
   //
   //--------------------------------------------------------------------------
 
+  /**
+   * Indicates whether the component is in a loading state.
+   *
+   * @internal
+   */
+  @Prop({ mutable: true, reflect: true }) complete = false;
+
   /** When `true`, displays smaller and appears to the left of the text. */
   @Prop({ reflect: true }) inline = false;
 
@@ -31,10 +38,16 @@ export class Loader implements LocalizedComponent {
    * Use `"indeterminate"` if finding actual progress value is impossible. Otherwise, use `"determinate"` to have the value indicate the progress or `"determinate-value"` to have the value label displayed along the progress.
    *
    */
-  @Prop({ reflect: true }) type: "indeterminate" | "determinate" | "determinate-value";
+  @Prop({ reflect: true }) type: "indeterminate" | "determinate" | "determinate-value" =
+    "indeterminate";
 
   /** The component's value. Valid only for `"determinate"` indicators. Percent complete of 100. */
   @Prop() value = 0;
+
+  @Watch("value")
+  valueChangeHandler(): void {
+    this.complete = this.type.startsWith("determinate") && this.value === 100;
+  }
 
   /** Text that displays under the component's indicator. */
   @Prop() text = "";
@@ -55,6 +68,10 @@ export class Loader implements LocalizedComponent {
     disconnectLocalized(this);
   }
 
+  componentWillLoad(): void {
+    requestAnimationFrame(() => this.valueChangeHandler());
+  }
+
   render(): VNode {
     const { el, inline, label, scale, text, type, value } = this;
 
@@ -63,25 +80,21 @@ export class Loader implements LocalizedComponent {
     const size = inline ? this.getInlineSize(scale) : this.getSize(scale);
     const radius = size * radiusRatio;
     const viewbox = `0 0 ${size} ${size}`;
-    const isDeterminate = type?.startsWith("determinate");
+    const isDeterminate = type.startsWith("determinate");
     const circumference = 2 * radius * Math.PI;
     const progress = (value / 100) * circumference;
     const remaining = circumference - progress;
     const valueNow = Math.floor(value);
-    const hostAttributes = {
-      "aria-valuenow": valueNow,
-      "aria-valuemin": 0,
-      "aria-valuemax": 100,
-      complete: valueNow === 100,
-    };
-    const svgAttributes = { r: radius, cx: size / 2, cy: size / 2 };
     const determinateStyle = { "stroke-dasharray": `${progress} ${remaining}` };
+
     return (
       <Host
         aria-label={label}
+        aria-valuemax={isDeterminate ? 100 : undefined}
+        aria-valuemin={isDeterminate ? 0 : undefined}
+        aria-valuenow={isDeterminate ? valueNow : undefined}
         id={id}
         role="progressbar"
-        {...(isDeterminate ? hostAttributes : {})}
       >
         <div class={CSS.loaderParts}>
           {[1, 2, 3].map((index) => (
@@ -92,10 +105,10 @@ export class Loader implements LocalizedComponent {
                 [CSS.loaderPartId(index)]: true,
               }}
               key={index}
+              style={isDeterminate && index === 3 ? determinateStyle : undefined}
               viewBox={viewbox}
-              {...(index === 3 && isDeterminate ? { style: determinateStyle } : {})}
             >
-              <circle {...svgAttributes} />
+              <circle cx={size / 2} cy={size / 2} r={radius} />
             </svg>
           ))}
           {isDeterminate && <div class={CSS.loaderPercentage}>{this.formatValue()}</div>}