+++ packages/calcite-ui-icons/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.32.0-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.32.0-next.2...@esri/calcite-ui-icons@3.32.0-next.3) (2024-09-12)
+
+### Features
+
+- add web adaptor ([#10286](https://github.com/Esri/calcite-design-system/issues/10286)) ([791fd78](https://github.com/Esri/calcite-design-system/commit/791fd782ac275c27e7a32b37b62e9c48b4a18f42))
+
 ## [3.32.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.32.0-next.1...@esri/calcite-ui-icons@3.32.0-next.2) (2024-09-11)
 
 ### Features